# simd mask applyOpacity, 10x faster