before:
`applyOpacity ....................... 1.797 ms      1.827 ms    ±0.049  x1000`
after:
`applyOpacity ....................... 0.165 ms      0.175 ms    ±0.017  x1000`